feat: flag duplicate handling
