+++ src/helpers/validate-flags.js
@@ -1,19 +1,13 @@
-import flagsError from './flags-error.js';
 import getFlags from './get-flags.js';
+import isFlagsAllowed from './is-flag-allowed.js';
+import isFlagsDuplicate from './is-flag-duplicate.js';
 
 const validateFlags = () => {
   const flags = getFlags();
-  console.log('flags: ', flags);
-  const requiredFlags = ['-c', '--config'];
-  const isFlagAllowed = requiredFlags.some((flag) =>
-    flags.includes(flag),
-  );
 
-  if (!isFlagAllowed) {
-    throw new flagsError(
-      'Stop, fellow! You need to enter valid options with flag -c or --config ',
-    );
-  }
+  isFlagsAllowed(flags);
+
+  isFlagsDuplicate(flags);
 };
 
 export default validateFlags;
